+++ stubs/config.full.js
@@ -493,6 +493,7 @@ module.exports = {
     },
     gridTemplateColumns: {
       none: 'none',
+      subgrid: 'subgrid',
       1: 'repeat(1, minmax(0, 1fr))',
       2: 'repeat(2, minmax(0, 1fr))',
       3: 'repeat(3, minmax(0, 1fr))',
@@ -508,6 +509,7 @@ module.exports = {
     },
     gridTemplateRows: {
       none: 'none',
+      subgrid: 'subgrid',
       1: 'repeat(1, minmax(0, 1fr))',
       2: 'repeat(2, minmax(0, 1fr))',
       3: 'repeat(3, minmax(0, 1fr))',